Refactor DTOs: use long IDs, remove validation, add short DTOs
Refactored DTOs to use long instead of int for IDs, removed data annotation validation attributes, and introduced lightweight "ForShortResultDto" classes for multiple entities. Updated service interfaces and replaced OrderForUpdateStatusDto with OrderForUpdateDto. Cleaned up usings and simplified DTO property initializations.

+++ SaraTort.Shared/DTOs/CakeOption/CakeOptionForResultDto.cs
@@ -2,8 +2,8 @@
 
 public class CakeOptionForResultDto
 {
-    public int Id { get; set; }
-    public int CakeId { get; set; }
+    public long Id { get; set; }
+    public long CakeId { get; set; }
     public double WeightInKg { get; set; }
     public decimal Price { get; set; }
     public int StockQuantity { get; set; }
